Refactor service discovery controller tests
Moved the test variables and supporting code to a
testDriver struct.
